+++ BilibiliMonitor/Models/VideoModel.cs
@@ -69,7 +69,7 @@ namespace BilibiliMonitor.Models
 
         public class Owner
         {
-            public int mid { get; set; }
+            public long mid { get; set; }
             public string name { get; set; }
             public string face { get; set; }
         }
